Implement heading with numbering
